# backend-service/doc/swagger.yaml
openapi: 3.0.3
info:
  title: Dell Technologies Resource Management Application Backend API Documentation
  description: >-
    This is the backend API documentation and specifcation for the Dell Technologies Resource Management Web Application, and it is based on the OpenAPI 3.0 specification.


    The API is a RESTful API service which we designed and built to allow us to perform CRUD operations on data stored on a MongoDB Server.


    This application is built using the following technolgoies:

    - [React.js](https://react.dev/)

    - [Node.js](https://nodejs.org/en)

    - [Express.js](https://expressjs.com/)

    - [MongoDB](https://www.mongodb.com/)

    - [Docker](https://www.docker.com/)

    - [Swagger.io](https://swagger.io/)


    This application is built to manage resources much easier 
    and more efficiently helping the organisation to keep 
    track and manage resources more easily. 


    It is built using the MERN stack and it is a full-stack 
    application which can be easily deployed with the help 
    of Docker deployment. We also utilised React.js as a 
    frontend framework which works well with Express.js and 
    Node.js as the backend server. For the database, we have
    used MongoDB, since it is a NoSQL database, it allows 
    for much more flexibility as well as scalability.
    Lastly, for the ease of testing and documentation, we 
    have utilised Swagger.io to document our API endpoints.


...
    below.


...

...
  contact:
...
  license:
...
servers:
...
paths:
  /admin/login:
    get:
      tags:
        - admin
      summary: Logs admin into the system
      description: Logs admin into the system
      operationId: adminLogin
      parameters:
        - name: credential
          in: query
          description: The username / email for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "authentication: successful operation"
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
            application/xml:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: username or password incorrect
        "500":
          description: "Error Logging In: Internal Server Error"
  /admin/logout:
    get:
      tags:
        - admin
      summary: Logs out current logged in admin session
      description: Logs out current logged in admin session
      operationId: adminLogout
      responses:
        "200":
          description: successful operation
        "500":
          description: "Error Logging Out: Internal Server Error"
  /admin/getAdmins:
    get:
      tags:
        - admin
      summary: Get all admin users
      description: Get all admin users
      operationId: getAllAdmin
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
        "401":
          description: No authorization token provided
        "403":
          description: Access forbidden
        "500":
          description: "Error Fetching Admins: Internal Server Error"
      security:
        - admin_auth_token: []
  /admin/createAdmin:
    post:
      tags:
        - admin
      summary: Create a new admin user
      description: Create a new admin user
      operationId: createAdmin
      requestBody:
        description: Create a new admin user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Admin"
          application/xml:
            schema:
              $ref: "#/components/schemas/Admin"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Admin"
        required: true
      security:
        - admin_auth_token: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
            application/xml:
              schema:
                $ref: "#/components/schemas/Admin"
        "400":
          description: Invalid input
        "401":
          description: No authorization token provided
        "422":
          description: Validation exception
  /admin/delete/{adminId}:
    delete:
      tags:
        - admin
      summary: Delete an admin by ID
      description: Delete an admin by ID
      operationId: deleteAdmin
      parameters:
        - name: adminId
          in: path
          description: ID of Admin to delete
          required: true
          schema:
            type: string
            description: MongoDB ObjectID as a hexadecimal string.
      security:
        - admin_auth_token: []
      responses:
        "200":
          description: successful operation
        "401":
          description: No authorization token provided
        "404":
          description: Admin not found
        "500":
          description: "Error Deleting Admin: Internal Server Error"
...
components:
  schemas:
    Admin:
      type: object
      properties:
        _id:
          type: string
          example: 668d5ba4a5f8eca9ff47283e
          description: MongoDB ObjectID as a hexadecimal string.
        username:
          type: string
          example: theAdmin
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: johndoe@example.com
        password:
          type: string
          example: "12345"
          description: Password is hashed before storing in the database.
        role:
          type: string
          example: admin
          enum:
            - admin
            - trainer
            - client
      xml:
        name: admin
...
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    admin_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    trainer_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    client_auth_token:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header